Factor out trees to support evaluation without training
