+++ brushfire-tree/build.sbt
@@ -1,0 +1,13 @@
+name := "brushfire-tree"
+
+libraryDependencies ++= {
+  import Deps._
+  Seq(
+    algebirdCore,
+    bonsai,
+    scalaTest,
+    scalaCheck
+  )
+}
+
+Publish.settings
