+++ brushfire-tree/src/main/scala/com/stripe/brushfire/AnnotatedTree.scala
