+++ brushfire-tree/src/main/scala/com/stripe/brushfire/ConfusionMatrix.scala
