+++ brushfire-tree/src/main/scala/com/stripe/brushfire/Predicate.scala
