+++ brushfire-tree/src/main/scala/com/stripe/brushfire/Reorder.scala
