+++ brushfire-tree/src/main/scala/com/stripe/brushfire/Split.scala
@@ -1,0 +1,18 @@
+package com.stripe.brushfire
+
+/** Candidate split for a tree node */
+case class Split[V, T](predicate: Predicate[V], leftDistribution: T, rightDistribution: T) {
+
+  def map[U](f: V => U): Split[U, T] =
+    Split(predicate.map(f), leftDistribution, rightDistribution)
+
+  /**
+   * Given a feature key, create a SplitNode from this Split.
+   *
+   * Note that the leaves of this node will likely need to be
+   * renumbered if this node is put into a larger tree.
+   */
+  def createSplitNode[K](feature: K): SplitNode[K, V, T, Unit] =
+    SplitNode(feature, predicate, LeafNode(0, leftDistribution), LeafNode(1, rightDistribution))
+}
+
