+++ brushfire-tree/src/main/scala/com/stripe/brushfire/TreeTraversal.scala
