+++ brushfire-tree/src/main/scala/com/stripe/brushfire/Voter.scala
@@ -80,3 +80,20 @@ object Voter {
 
   private def mode[L, M: Ordering](m: Map[L, M]): L = m.maxBy { _._2 }._1
 }
+
+// Backwards compat.
+
+object SoftVoter {
+  def apply[L, M: Numeric]() = Voter.soft[L, M]
+}
+
+object ModeVoter {
+  def apply[L, M: Ordering]() = Voter.mode[L, M]
+}
+
+object ThresholdVoter {
+  import Voter.FrequencyVoter
+
+  def apply[M](threshold: Double, voter: FrequencyVoter[Boolean, M]) =
+    Voter.threshold(threshold, voter)
+}
