+++ brushfire-tree/src/main/scala/com/stripe/brushfire/package.scala
